Add memory and # of core options to JVM_OPTS
This sets, based on any limits placed on the docker container, an
Xmx (max heap size) and some options that depend on the # of cores
available. Without this, the jvm will make its own calculations based on
the total ram and # of cores of the host machine, which may be beyond
its limits. These settings should prevent the jvm from being killed by
docker for exceeding those limits.

This also includes an option to have the jvm kill itself if it runs out
of heap, instead of lingering and throwing errors.

Implements ENG-249: https://farmlogs.atlassian.net/browse/ENG-249