Disable fast throw stack elision
There are cases where the HotSpot VM will recompile methods that throw
the same exception often, replacing the exception with a pregenerated
one that does not include a stack trace. When that happens, we don't get
a stack trace reported to Sentry, which makes debugging the root cause
difficult. This makes the JVM generate stack traces as the default
behavior at the cost of a slight performance degradation in cases where
the same exception is thrown lots of times (which generally happens when
exceptions are used for flow control).

ENG-284: https://farmlogs.atlassian.net/browse/ENG-284